+++ example/client/src/main.ts
@@ -1,5 +1,5 @@
 import './style.css'
 
-// import './fetch'
+import './fetch'
 // import './fn'
 import './treaty'